secret-wrapper: move the underlying dir while keeping previous one
