+++ images/entrypoint-wrapper/Dockerfile
@@ -1,0 +1,5 @@
+FROM centos:8
+LABEL maintainer="bbarcaro@redhat.com"
+
+ADD entrypoint-wrapper /usr/bin/entrypoint-wrapper
+ENTRYPOINT ["/usr/bin/entrypoint-wrapper"]
